pyproject.toml
[tool.poetry]
name = "yawt"
version = "0.1.0"
description = "YAWT (Yet Another Whisper-based Transcriber) is a transcription tool that leverages OpenAI's Whisper model to deliver accurate and efficient audio-to-text conversion."
authors = ["Yaniv Golan <yaniv@golan.name>"]
packages = [{ include = "yawt", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11.4"
torch = "^2.1.0"
transformers = ">=4.35.0"
tqdm = ">=4.66.0"
python-dotenv = ">=1.0.0"
ffmpeg-python = ">=0.2.0"
numpy = ">=1.21.0"
requests = ">=2.31.0"
requests-toolbelt = ">=0.10.1"
accelerate = ">=0.26.0"
srt = ">=3.0.0"
pyyaml = ">=6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
yawt = "yawt.main:main"